Modify 'frame_rate' to be integer in SimulationSettings.

+++ SimulationSettings.pde
@@ -2,7 +2,7 @@ public class SimulationSettings
 {
   public float FeedRate = 0; // Default feed rate is '0'.
   public float KillRate = 0; // Default kill rate is '0'.
-  public float FrameRate = 60; // Default frame rate is '60'.
+  public int FrameRate = 60; // Default frame rate is '60'.
   
   SimulationSettings()
   {
